+++ sff/structs.py
@@ -195,7 +195,8 @@ class SettingItem(NamedTuple):
 class Settings(Enum):
     ADVANCED_MODE = SettingItem("advanced_mode", "Advanced Mode", False, bool)
     HUBCAP_KEY = SettingItem("morrenus_key", "Hubcap API Key", True, str)
-    RYUU_KEY = SettingItem("ryuu_key", "Ryuu API Key", True, str)
+    RYUU_KEY = SettingItem("ryuu_key", "Ryuu Reseller Key", True, str)
+    RYUU_API_KEY = SettingItem("ryuu_api_key", "Ryuu API Key (premium)", True, str)
     STEAM_PATH = SettingItem(
         "steam_path", "Steam Installation Path", False, SettingCustomTypes.DIR
     )